+++ docs/source/changelog.rst
@@ -9,9 +9,9 @@ Changelog
 Version (dev)
 -------------
 
-* Backwards-incompatible change: Change type for FGD keyvalue options to :class:`~srctools.fgd.KVOption`.
 * Add :mod:`srctools.sndscape`, for parsing soundscape scripts.
 * Changed `PackList.load_choreo_scenes() <srctools.packlist.PackList.load_choreo_scenes>` to warn if ``scenes.image`` is not found.
+* Change type for FGD keyvalue options to :class:`~srctools.fgd.KVOption`. Previous tuple attributes are deprecated.
 * Add support for ``orientedwidthheight()``, ``orientedwidthheighthalf()`` and ``clusteredlight()`` Strata FGD helpers.
 * Add support for descriptions in spawnflags - displayed in Strata Source.
 * Add ``noinherit()`` FGD helper, for indicating values to avoid inheriting from base classes.